move userlist to model folder
