+++ StudyLink/src/model/UserList.java
@@ -1,6 +1,4 @@
-package view;
-
-import model.*;
+package model;
 
 import java.util.HashMap;
 import java.util.Map;
